pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pokemon-stock-checker"
version = "0.1.0"
description = "Local Pokemon TCG stock checker with safe polling and alerts"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Codex" }]
dependencies = [
  "pydantic>=2.6,<3",
  "PyYAML>=6.0",
  "requests>=2.31,<3",
  "playwright>=1.50,<2",
  "rich>=13.7,<14",
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
  "pytest-cov>=5.0,<6",
]

[project.scripts]
stockcheck = "stockcheck.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
